server/requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
pydantic
pydantic-settings
slowapi
loguru
bcrypt==4.0.1